Migrate intersects geometry
